Move cephlcm.controller to cephlcm_controller
