+++ backend/controller/cephlcm_controller/inventory.py
@@ -19,7 +19,7 @@ from cephlcm_common import playbook_plugin
 from cephlcm_common import plugins
 from cephlcm_common import wrappers
 from cephlcm_common.models import generic
-from cephlcm.controller import exceptions
+from cephlcm_controller import exceptions
 
 
 CONF = config.make_controller_config()
