+++ backend/controller/setup.cfg
@@ -1,13 +1,13 @@
 [metadata]
-name = cephlcm
+name = cephlcm_controller
 version = 0.1
-summary = Ceph lifecycle management
+summary = Ceph lifecycle management controller service
 maintainer = Mirantis
 mainteiner-email = sarkhipov@mirantis.com
 author = Mirantis
 author-email = sarkhipov@mirantis.com
 license = Apache2
-requires-python = >=2.7
+requires-python = >=3.4
 requires-dist =
   setuptools
 home-page = https://github.com/Mirantis/ceph-lcm
@@ -17,23 +17,15 @@ classifier =
   Intended Audience :: System Administrators
   Operating System :: POSIX :: Linux
   Programming Language :: Python
-  Programming Language :: Python :: 2.7
+  Programming Language :: Python :: 3.4
+  Programming Language :: Python :: 3.5
 
 [files]
 packages =
-  cephlcm
-  cephlcmlib
-
-package_data =
-  cephlcm =
-    configs/*.toml
-
-[entry_points]
-cephlcm.alerts =
-  email = plugins.alerts.email:alert
-
+  cephlcm_controller
 console_scripts =
-  cephlcm = cephlcmlib.cli:cli
+  cephlcm-controller = cephlcm_controller.daemon:main
+  cephlcm-inventory = cephlcm_controller.inventory:main
 
 [global]
 setup-hooks =
@@ -45,8 +37,5 @@ warnerrors = true
 [flake8]
 exclude = build,dist,docs/conf.py,*.egg-info,.eggs,.cache,.tox,plugins/playbook/deploy_cluster/cephlcm_deploy_cluster/playbooks
 inline-quotes = "
-application-import-names = cephlcm_api,cephlcm_common,cephlcm_controller,cephlcmlib
+application-import-names = cephlcm_api,cephlcm_common,cephlcm_controller
 import-order-style = smarkets
-
-[tool:pytest]
-addopts = -vvv -l
