+++ backend/controller/setup.py
@@ -1,0 +1,18 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+
+
+import setuptools
+
+
+try:
+    import multiprocessing
+    assert multiprocessing
+except ImportError:
+    pass
+
+
+setuptools.setup(
+    setup_requires=["pbr>=1.10"],
+    pbr=True
+)
