+++ tests/controller/test_taskpool.py
@@ -11,7 +11,7 @@ import pytest
 
 from cephlcm_common import config
 from cephlcm_common.models import task
-from cephlcm.controller import taskpool
+from cephlcm_controller import taskpool
 
 
 CONF = config.make_controller_config()
